remove loq plans (moved these to instrument)
