+++ src/ibex_bluesky_core/devices/sans/__init__.py
@@ -1,0 +1,1 @@
+"""Devices specific to SANS beamlines."""
