+++ src/ibex_bluesky_core/plans/muons/__init__.py
@@ -1,1 +1,1 @@
-"""Muon plans and helpers."""
+"""Plans specific to Muon beamlines."""
