+++ src/ibex_bluesky_core/plans/sans/__init__.py
@@ -1,1 +1,1 @@
-"""SANS Plans and helpers."""
+"""Plans specific to SANS beamlines."""
